Add comments system using GitHub Discussions via giscus
Comments are enabled by default, and can be disabled by adding
`disable_comments: true` to the front matter of a page:

```markdown
---
disable_comments: true
---

Page content [...]
```

This PR disables comments on the main home page and "About" page.

This largely uses the approach described in the documentation at
https://squidfunk.github.io/mkdocs-material/setup/adding-a-comment-system/

However, as mkdocs-material's privacy plugin is not compatible with
Giscus, this uses the alternative implementation approach described in
mkdocs-material upstream PR 6502.